% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml.lrt.R
\name{asreml.lrt}
\alias{asreml.lrt}
\title{Calculate the LRT test for model compare.}
\usage{
asreml.lrt(m1 = NULL, m2 = NULL)
}
\arguments{
\item{m1}{The asreml object}

\item{m2}{The asreml object}
}
\value{
The LRT test for \code{m1} and \code{m2}
}
\description{
\code{asreml.lrt} can Calculate the LRT test for model compare.
}
\examples{
library(asreml)
library(vsnc)
data(oats,package = "asreml")
head(oats)
m1 <- asreml(yield ~ 1, random=~Nitrogen*Variety, data=oats)
m2 <- asreml(yield ~ 1, random=~Nitrogen*Variety+Blocks, data=oats)
asreml.lrt(m1,m2)
}
\seealso{
\code{\link{model.comp}}
}
\author{
Dave <Dave@vsni.co.uk>
}
