% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTdata.R
\docType{data}
\name{BTdata}
\alias{BTdata}
\title{Blue Tit Data for a Quantitative Genetic Experiment}
\format{\describe{A data frame with 828 rows and 7 columns.
\item{tarsus}{with variables tarsus length.}
\item{back}{colour.}
\item{animal}{individuals.}
\item{dam}{The mother of each is also recorded.}
\item{fosternest}{with the foster nest.}
\item{hatchdate}{first egg in each nest hatched.}
\item{sex}{sex.} 
}}
\usage{
BTdata
}
\description{
Blue Tit (Cyanistes caeruleus) Data for a Quantitative Genetic Experiment
}
\references{
MCMCglmm package, Hadfield, J.D. et. al. 2007 Journal of Evolutionary Biology 20 549-557
}
\seealso{
\code{\link{BTped}}
}
\keyword{datasets}
