% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTped.R
\docType{data}
\name{BTped}
\alias{BTped}
\title{Blue Tit Pedigree for BTdata}
\format{\describe{a data frame with 1040 rows and 3 columns, with individual identifier (animal) mother identifier (dam) and father identifier (sire). The first 212 rows are the parents of the 828 offspring from 106 full-sibling families. Parents are assumed to be unrelated to each other and have NA's in the dam and sire column.
\item{animal}{}
\item{dam}{}
\item{sire}{}
}}
\usage{
BTped
}
\description{
Blue Tit (Cyanistes caeruleus) Pedigree
}
\references{
MCMCglmm package, Hadfield, J.D. et. al. 2007 Journal of Evolutionary Biology 20 549-557
}
\seealso{
\code{\link{BTdata}}
}
\keyword{datasets}
