% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lettuce.R
\docType{data}
\name{Lettuce}
\alias{Lettuce}
\title{Effects of nitrogen, phosphorus and potasium on lettuce seedlings}
\format{\describe{A data frame with 32 observations on the following six variables:

\item{Replicate}{a factor specifying the replicates.}
\item{Block}{a factor specifying the blocks.}
\item{N}{a factor specifying the amount of nitrogen fertiliser.}
\item{P}{a factor specifying the amount of phosphorus fertiliser.}
\item{K}{a factor specifying the amount of potasium fertiliser.}
\item{Yield}{a numerical vector recording the yields.}
}}
\source{
Cochran, W.G. & Cox, G.M. (1957). Experimental Designs (2nd edition), page 196.
}
\usage{
Lettuce
}
\description{
The Lettuce dataset contains data from an experiment growing lettuce seedings. There are three treatment factors: nitrogen (N), phosphorus (P) and potash (K). It was arranged in three replicates, each containing three blocks, and a different set of contrasts of the N:P:K interaction was confounded with blocks within each replicate.
}
\keyword{datasets}
