% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.comp.R
\name{model.comp}
\alias{model.comp}
\title{Calculate the AIC and BIC}
\usage{
model.comp(m1 = NULL, m2 = NULL)
}
\arguments{
\item{m1}{The asreml object}

\item{m2}{The asreml object}
}
\value{
The AIC and BIC of \code{m1} and \code{m2}
}
\description{
Calculate the AIC and BIC
}
\examples{
library(asreml)
library(vsnc)
data(oats,package = "asreml")
head(oats)
m1 <- asreml(yield ~ 1, random=~Nitrogen*Variety, data=oats)
m2 <- asreml(yield ~ 1, random=~Nitrogen*Variety+Blocks, data=oats)
m3 <- asreml(yield ~ 1, random=~Nitrogen*Variety+Blocks/Wplots, data=oats)
model.comp(m1)
model.comp(m3,m2) 
}
\seealso{
\code{\link{asreml.lrt}}
}
\author{
Dave <Dave@vsni.co.uk>
}
