% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Octane.R
\docType{data}
\name{Octane}
\alias{Octane}
\title{Effect effect of different additives on the octane level of gasoline}
\format{\describe{A data frame with 20 observations on the following two variables:
\item{Octane}{a numerical vector recording the octane level of each gasoline sample.}
\item{Gasoline}{a factor specifying the additive used in each sample.}
}}
\source{
John, P.W.M. (1971) Statistical Design and Analysis of Experiments, page 46.
}
\usage{
Octane
}
\description{
The Octane dataset contains data from an experiment to study the effect of different additives on the octane level of gasoline. There were five types of gasoline (A-E), and four observations on each.
}
\keyword{datasets}
