% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin}
\alias{pin}
\title{Calculate the Heritibality and Genetic Correlation}
\usage{
pin(object, transform)
}
\arguments{
\item{object}{It is the asreml object}

\item{transform}{It is the formula of calculation}
}
\description{
This is the function that can calculate the Heritibality and Genetic Correlation
}
\examples{
library(asreml)
data(harvey,package = "asreml")
head(harvey)
ainv <- asreml.Ainverse(harvey[,1:3])$ginv
mod.asr <- asreml(y1 ~ Line, random = ~ ped(Calf), ginverse = list(Calf = ainv),data=harvey)
summary(mod.asr)$varcomp
pin(mod.asr,h2 ~ V1/(V1+V2))
}
\seealso{
\code{\link{model.comp}} , \code{\link{asreml.lrt}}
}
\author{
Dave <Dave@vsni.co.uk>
}
