% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plankton.R
\docType{data}
\name{Plankton}
\alias{Plankton}
\title{Abundance of types of plankton}
\format{\describe{A data frame with 48 observations on the following three variables:

\item{Haul}{a factor specifying the type of plankton.}
\item{Type}{a factor specifying whether fumigants have been to control nematodes. It has two levels, Not fumigated and fumigated.}
\item{Number}{a numerical vector recording the number of plankton.}
}}
\source{
Snedecor, G.W. & Cochran, W.G. (1967) Statistical Methods (6th Edition), page 329.
}
\usage{
Plankton
}
\description{
The Plankton dataset contains data from a study of plankton numbers. Four types of plankton were sampled in 12 hauls. The hauls can be regarded as blocks. The types of plankton are the treatments.
}
\keyword{datasets}
