% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riceMV.R
\docType{data}
\name{riceMV}
\alias{riceMV}
\title{Effects of nitrogen fertiliser on yields of varieties of rice}
\format{\describe{A data frame with 54 observations on the following six variables:
 \item{Pair}{a factor}
 \item{Run}{a factor}
 \item{Variety}{a factor specifying the variety grown on each plot.}
 \item{yc}{a variate}
 \item{ye}{a variate}
 \item{syc}{a variate}
 \item{sye}{a variate}
 }}
\source{
R package:asreml
}
\usage{
riceMV
}
\description{
The Rice dataset contains data from an experiment that studied 
the effect of three levels of nitrogen fertilizer on the yields 
of six varieties of rice.
}
\keyword{datasets}
