% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsnc.R
\name{vsnc}
\alias{vsnc}
\title{A package that provide data and analysis code in Agri}
\description{
That contains plant breeding, forest breeding, animal breeding etc.
It also have some functions that assist asreml package.
}
\author{
Dave<Dave@vsni.co.uk>
}
