% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wine.R
\docType{data}
\name{Wine}
\alias{Wine}
\title{Percentage of alcohol in wine}
\format{\describe{A data frame with 15 observations on the following two variables:

\item{wine}{a factor specifying the type of wine, labelled A to E, measured in each sample.}
\item{alcohol}{a numerical vector recording the percentage of alcohol.}
}}
\source{
VSN International (2014). Genstat for Windows 17th Edition. VSN International, Hemel Hempstead, UK. Web page: Genstat.co.uk
}
\usage{
Wine
}
\description{
The Wine dataset contains data from an experiment to assess the percentage of alcohol by volume in five types of wine. Three bottles of each type were tested in the laboratory in a random order.
}
\keyword{datasets}
